pyproject.toml
[project]
name = "pw"
version = "0.1.0"
description = "Personal website"
readme = "README.md"
requires-python = ">=3.12"
license = { file = "LICENSE"}
dependencies = [
    "coltrane<1",
    "django-htmx>=1.21.0",
    "mistune>=3.0.2",
    "msgspec",
    "pytailwindcss>=0.2.0",
]

[tool.uv.sources]
msgspec = { git = "https://github.com/jcrist/msgspec.git", rev = "main" }
coltrane = { git = "https://github.com/Tobi-De/coltrane", rev = "main" }
